pyproject.toml
[project]
name = "purplerepo"
version = "0.1.0"
description = "🛡️⚔️ Curated GitHub repos for Defensive & Offensive Cyber Tradecraft"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0",
    "GitPython>=3.1.0",
    "loguru>=0.7.0",
]

[dependency-groups]
dev = [
    "pre-commit",
    "pre-commit-uv"
]